Relocate Tile tyoe definition to commonTypes
